Update OpenCV post
This updates the OpenCV post to automatically detect most of the
variables, as well as the system prefix. This is based on the recent
work that was done in getting OpenCV running on Travis, which
required automatically determining where all of the locations were,
as you can't depend on them being consistent with Travis CI.

This also adds a note about the outdated version of CMake on older
versions of Ubuntu.