Started working on post about securing relations in DRF
This is going to be focused on the browsable API, since that's usually where people first realize their relations are not locked down. But the ultimate goal will be to explain how to lock down what is allowed by your related fields, since that's a very common issue and a lot of people forget to do it.